pyproject.toml
[tool.poetry]
name = "shortify"
version = "0.1.0"
description = ""
authors = ["Matthew Cochrane <matthew.cochrane.eng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.23.0"
load-dotenv = "^0.1.0"
newspaper3k = "^0.2.8"
ratelimit = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
